+++ app/views.py
@@ -143,7 +143,7 @@ def saveIncomeView(request):
             quantity = len(data)
         for i in data:
             try:
-                a = int(i)
+                a = float(i)
                 massa += a
             except:
                 return JsonResponse({'data': 'error'})
